ExpenseController.java - added verification to check if user is in group

+++ evenly_backend/src/main/java/com/evenly/controller/ExpenseController.java
@@ -4,8 +4,11 @@ import com.evenly.Utility.DivideUtility;
 import com.evenly.dto.EqualExpenseCreateRequestDTO;
 import com.evenly.dto.ExpenseCreateResponseDTO;
 import com.evenly.entity.Expense;
+import com.evenly.exception.InvalidCredentialException;
 import com.evenly.service.ExpenseService;
 import com.evenly.service.ExpenseShareService;
+import com.evenly.service.GroupMemberService;
+import com.evenly.service.JwtService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,11 +30,22 @@ public class ExpenseController {
     @Autowired
     private ExpenseShareService expenseShareService;
 
+    @Autowired
+    private GroupMemberService groupMemberService;
+
+    @Autowired
+    private JwtService jwtService;
+
     @PostMapping("/equal")
     @Operation(
             summary = "Create an expense for an user in a group"
     )
-    public ResponseEntity<ExpenseCreateResponseDTO> addExpenseEqual(@RequestBody EqualExpenseCreateRequestDTO expense) {
+    public ResponseEntity<ExpenseCreateResponseDTO> addExpenseEqual(@RequestHeader("Authorization") String authorizationHeader, @RequestBody EqualExpenseCreateRequestDTO expense) {
+        String userId = jwtService.extractUsername(authorizationHeader.substring(7));
+        if (!groupMemberService.isMember(expense.getGroupId(), userId)) {
+            throw new InvalidCredentialException("Invalid credential.");
+        }
+
         Expense createdExpense = expenseService.addExpense(expense);
         ExpenseCreateResponseDTO responseDTO = new ExpenseCreateResponseDTO();
         responseDTO.setId(createdExpense.getId());
@@ -51,7 +65,13 @@ public class ExpenseController {
     }
 
     @DeleteMapping
-    public ResponseEntity<HttpStatus> deleteExpense(@RequestParam("expenseId") String expenseId) {
+    public ResponseEntity<HttpStatus> deleteExpense(@RequestHeader("Authorization") String authorizationHeader, @RequestParam("expenseId") String expenseId) {
+        String userId = jwtService.extractUsername(authorizationHeader.substring(7));
+        Expense expenseToBeDeleted = expenseService.getExpense(expenseId);
+        if (expenseToBeDeleted == null || !groupMemberService.isMember(expenseToBeDeleted.getGroupId(), userId)) {
+            throw new InvalidCredentialException("Invalid credential.");
+        }
+
         expenseService.deleteExpense(expenseId);
         expenseShareService.delete(expenseId);
         return new ResponseEntity<>(HttpStatus.OK);
